# api-gateway/docs/swagger.yaml
basePath: /
definitions:
...
  handler.peerOtcOfferReq:
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: EXBanka Banking Microservices API Gateway. All endpoints are served
    under /api/v3 — v1 and v2 have been retired (plan E, 2026-04-27). Future versions
    (v4+) will be added as separate explicit router files with no transparent fallback.
    See api-gateway/internal/router/router_versioning.md.
  title: EXBanka API
  version: "3.0"
paths:
...
  /api/v3/cross-bank-protocol/negotiations:
    post:
      consumes:
      - application/json
      description: Inbound from a peer bank's SI-TX layer. Authenticated via PeerAuth
        (X-Api-Key or HMAC). Persists a peer_otc_negotiations row keyed on (peer_bank_code,
        foreign_id). buyerId.routingNumber MUST match the authenticated peer's routing
        and sellerId.routingNumber MUST equal this bank.
      parameters:
      - description: SI-TX OtcOffer wire shape
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.peerOtcOfferReq'
      produces:
      - application/json
      responses:
        "201":
          description: ForeignBankId of the newly-created negotiation
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      summary: 'Peer-to-peer: create OTC option negotiation'
      tags:
      - PeerOTC
  /api/v3/cross-bank-protocol/negotiations/{rid}/{id}:
    delete:
      description: Inbound from a peer bank. Soft-cancels the negotiation row (status
        → cancelled). The row is preserved for audit per SI-TX §3.5.
      parameters:
      - description: routing number
        in: path
        name: rid
        required: true
        type: integer
      - description: foreign negotiation id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: 'Peer-to-peer: cancel an OTC negotiation'
      tags:
      - PeerOTC
    get:
      description: Inbound from a peer bank. Returns the full SI-TX OtcNegotiation
        record for the (peer_bank_code, foreign_id) pair.
      parameters:
      - description: routing number
        in: path
        name: rid
        required: true
        type: integer
      - description: foreign negotiation id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: 'Peer-to-peer: read an OTC negotiation'
      tags:
      - PeerOTC
    put:
      consumes:
      - application/json
      description: Inbound from a peer bank. Updates the offer JSON on a peer_otc_negotiations
        row. Only the peer that created the row (matched by peer_bank_code) can update
        it — peer-auth + (peer_bank_code, negotiation_id) lookup combo enforces this.
      parameters:
      - description: routing number (this bank's, that issued the foreign_id)
        in: path
        name: rid
        required: true
        type: integer
      - description: foreign negotiation id
        in: path
        name: id
        required: true
        type: string
      - description: new SI-TX OtcOffer terms
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.peerOtcOfferReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "409":
          description: out of turn or negotiation closed (SI-TX §3.3)
          schema:
            additionalProperties: true
            type: object
      summary: 'Peer-to-peer: counter-offer on an existing OTC negotiation'
      tags:
      - PeerOTC
  /api/v3/cross-bank-protocol/negotiations/{rid}/{id}/accept:
    get:
      description: Inbound from the buyer's bank. The seller's bank composes the 4-posting
        NEW_TX (buyer-debit premium, seller-credit premium, seller-debit option asset,
        buyer-credit option asset) and dispatches via PeerTxService. SI-TX §3.6 specifies
        GET semantics.
      parameters:
      - description: routing number
        in: path
        name: rid
        required: true
        type: integer
      - description: foreign negotiation id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: transactionId + status
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: 'Peer-to-peer: accept an OTC negotiation (triggers option-formation
        SI-TX)'
      tags:
      - PeerOTC
  /api/v3/cross-bank-protocol/public-option-offers:
    get:
      description: Phase 6 cross-bank discovery. Returns this bank's OPEN, undirected
        option listings as PeerPublicOptionOffer rows in SI-TX shape. Auth via X-Api-Key
        (PeerAuth); X-Bank-Code is stamped into peer_bank_code so privately-targeted
        listings are filtered per-caller.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "501":
          description: OTCOfferReader not wired
          schema:
            additionalProperties: true
            type: object
      summary: Peer-facing list of OPEN OTC option listings on this bank
      tags:
      - PeerOTC
  /api/v3/cross-bank-protocol/public-stock:
    get:
      description: Inbound from a peer bank. Returns this bank's holdings flagged
        public_quantity > 0 — the candidate OTC sellers a discovering bank can negotiate
        against. SI-TX §3.2.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: 'Peer-to-peer: list public stock holdings'
      tags:
      - PeerOTC
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"